requirements.txt
asgiref==3.6.0
Django==4.1.5
django-admin-interface==0.24.2
django-colorfield==0.8.0
docstring-to-markdown==0.11
jedi==0.18.2
parso==0.8.3
Pillow==9.4.0
pluggy==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.0
python-slugify==7.0.0
sqlparse==0.4.3
text-unidecode==1.3
ujson==5.6.0
waitress==2.1.2
whitenoise==6.3.0
numpy==1.24.1
pandas==1.5.2
plotly==5.11.0
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
tenacity==8.1.0
